# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRON Car Rental API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type CarTransmission =
  (typeof CarTransmission)[keyof typeof CarTransmission];

export const CarTransmission = {
  automatic: "automatic",
  manual: "manual",
} as const;

export type CarFuelType = (typeof CarFuelType)[keyof typeof CarFuelType];

export const CarFuelType = {
  petrol: "petrol",
  diesel: "diesel",
  electric: "electric",
  hybrid: "hybrid",
} as const;

export interface Car {
  id: number;
  make: string;
  model: string;
  year: number;
  color: string;
  category: string;
  transmission: CarTransmission;
  fuelType: CarFuelType;
  seats: number;
  bags: number;
  pricePerDay: number;
  available: boolean;
  imageUrl: string | null;
  features: string[];
  description: string | null;
  createdAt: string;
}

export type CreateCarInputTransmission =
  (typeof CreateCarInputTransmission)[keyof typeof CreateCarInputTransmission];

export const CreateCarInputTransmission = {
  automatic: "automatic",
  manual: "manual",
} as const;

export type CreateCarInputFuelType =
  (typeof CreateCarInputFuelType)[keyof typeof CreateCarInputFuelType];

export const CreateCarInputFuelType = {
  petrol: "petrol",
  diesel: "diesel",
  electric: "electric",
  hybrid: "hybrid",
} as const;

export interface CreateCarInput {
  make: string;
  model: string;
  year: number;
  color: string;
  category: string;
  transmission: CreateCarInputTransmission;
  fuelType: CreateCarInputFuelType;
  seats: number;
  bags: number;
  pricePerDay: number;
  available: boolean;
  imageUrl?: string | null;
  features?: string[];
  description?: string | null;
}

export interface Location {
  id: number;
  name: string;
  address: string;
  city: string;
  country: string;
  isAirport: boolean;
  active: boolean;
}

export interface CreateLocationInput {
  name: string;
  address: string;
  city: string;
  country: string;
  isAirport: boolean;
  active: boolean;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface Booking {
  id: number;
  carId: number;
  car?: Car | null;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  pickupLocationId: number;
  dropoffLocationId: number;
  pickupLocation?: Location | null;
  dropoffLocation?: Location | null;
  pickupDate: string;
  dropoffDate: string;
  totalDays: number;
  totalPrice: number;
  status: BookingStatus;
  notes: string | null;
  createdAt: string;
}

export interface CreateBookingInput {
  carId: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  pickupLocationId: number;
  dropoffLocationId: number;
  pickupDate: string;
  dropoffDate: string;
  notes?: string | null;
}

export type UpdateBookingInputStatus =
  (typeof UpdateBookingInputStatus)[keyof typeof UpdateBookingInputStatus];

export const UpdateBookingInputStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface UpdateBookingInput {
  status: UpdateBookingInputStatus;
  notes?: string | null;
  customerName?: string;
  customerEmail?: string;
  customerPhone?: string;
  pickupLocationId?: number;
  dropoffLocationId?: number;
  pickupDate?: string;
  dropoffDate?: string;
}

export interface AdminStats {
  totalCars: number;
  availableCars: number;
  totalBookings: number;
  pendingBookings: number;
  confirmedBookings: number;
  completedBookings: number;
  cancelledBookings: number;
  totalRevenue: number;
  recentBookings: Booking[];
}

export type ListCarsParams = {
  available?: boolean;
  category?: string;
};

export type ListBookingsParams = {
  status?: ListBookingsStatus;
  carId?: number;
};

export type ListBookingsStatus =
  (typeof ListBookingsStatus)[keyof typeof ListBookingsStatus];

export const ListBookingsStatus = {
  pending: "pending",
  confirmed: "confirmed",
  cancelled: "cancelled",
  completed: "completed",
} as const;