# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RRON Car Rental API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cars
    description: Car fleet management
  - name: bookings
    description: Booking management
  - name: locations
    description: Pickup/dropoff locations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /cars:
    get:
      operationId: listCars
      tags: [cars]
      summary: List all cars
      parameters:
        - name: available
          in: query
          schema:
            type: boolean
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
    post:
      operationId: createCar
      tags: [cars]
      summary: Create a new car
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarInput"
      responses:
        "201":
          description: Created car
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"

  /cars/{id}:
    get:
      operationId: getCar
      tags: [cars]
      summary: Get a car by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Car details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "404":
          description: Car not found
    put:
      operationId: updateCar
      tags: [cars]
      summary: Update a car
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarInput"
      responses:
        "200":
          description: Updated car
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
    delete:
      operationId: deleteCar
      tags: [cars]
      summary: Delete a car
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List all pickup/dropoff locations
      responses:
        "200":
          description: List of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Location"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationInput"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"

  /locations/{id}:
    put:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationInput"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - name: carId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    put:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingInput"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Admin statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    Car:
      type: object
      properties:
        id:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        category:
          type: string
        transmission:
          type: string
          enum: [automatic, manual]
        fuelType:
          type: string
          enum: [petrol, diesel, electric, hybrid]
        seats:
          type: integer
        bags:
          type: integer
        pricePerDay:
          type: number
        available:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        features:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - make
        - model
        - year
        - color
        - category
        - transmission
        - fuelType
        - seats
        - bags
        - pricePerDay
        - available
        - imageUrl
        - features
        - description
        - createdAt

    CreateCarInput:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        category:
          type: string
        transmission:
          type: string
          enum: [automatic, manual]
        fuelType:
          type: string
          enum: [petrol, diesel, electric, hybrid]
        seats:
          type: integer
        bags:
          type: integer
        pricePerDay:
          type: number
        available:
          type: boolean
        imageUrl:
          type: string
          nullable: true
        features:
          type: array
          items:
            type: string
        description:
          type: string
          nullable: true
      required:
        - make
        - model
        - year
        - color
        - category
        - transmission
        - fuelType
        - seats
        - bags
        - pricePerDay
        - available

    Location:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        isAirport:
          type: boolean
        active:
          type: boolean
      required:
        - id
        - name
        - address
        - city
        - country
        - isAirport
        - active

    CreateLocationInput:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        isAirport:
          type: boolean
        active:
          type: boolean
      required:
        - name
        - address
        - city
        - country
        - isAirport
        - active

    Booking:
      type: object
      properties:
        id:
          type: integer
        carId:
          type: integer
        car:
          $ref: "#/components/schemas/Car"
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        pickupLocationId:
          type: integer
        dropoffLocationId:
          type: integer
        pickupLocation:
          $ref: "#/components/schemas/Location"
          nullable: true
        dropoffLocation:
          $ref: "#/components/schemas/Location"
          nullable: true
        pickupDate:
          type: string
        dropoffDate:
          type: string
        totalDays:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - carId
        - customerName
        - customerEmail
        - customerPhone
        - pickupLocationId
        - dropoffLocationId
        - pickupDate
        - dropoffDate
        - totalDays
        - totalPrice
        - status
        - notes
        - createdAt

    CreateBookingInput:
      type: object
      properties:
        carId:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        pickupLocationId:
          type: integer
        dropoffLocationId:
          type: integer
        pickupDate:
          type: string
        dropoffDate:
          type: string
        notes:
          type: string
          nullable: true
      required:
        - carId
        - customerName
        - customerEmail
        - customerPhone
        - pickupLocationId
        - dropoffLocationId
        - pickupDate
        - dropoffDate

    UpdateBookingInput:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: string
          nullable: true
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        pickupLocationId:
          type: integer
        dropoffLocationId:
          type: integer
        pickupDate:
          type: string
        dropoffDate:
          type: string
      required:
        - status

    AdminStats:
      type: object
      properties:
        totalCars:
          type: integer
        availableCars:
          type: integer
        totalBookings:
          type: integer
        pendingBookings:
          type: integer
        confirmedBookings:
          type: integer
        completedBookings:
          type: integer
        cancelledBookings:
          type: integer
        totalRevenue:
          type: number
        recentBookings:
          type: array
          items:
            $ref: "#/components/schemas/Booking"
      required:
        - totalCars
        - availableCars
        - totalBookings
        - pendingBookings
        - confirmedBookings
        - completedBookings
        - cancelledBookings
        - totalRevenue
        - recentBookings