# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRON Car Rental API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";
import type { Car } from "./car";
import type { Location } from "./location";

export interface Booking {
  id: number;
  carId: number;
  car?: Car | null;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  pickupLocationId: number;
  dropoffLocationId: number;
  pickupLocation?: Location | null;
  dropoffLocation?: Location | null;
  pickupDate: string;
  dropoffDate: string;
  totalDays: number;
  totalPrice: number;
  status: BookingStatus;
  notes: string | null;
  createdAt: string;
}