# lib/api-zod/src/generated/types/car.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRON Car Rental API
 * OpenAPI spec version: 0.1.0
 */
import type { CarFuelType } from "./carFuelType";
import type { CarTransmission } from "./carTransmission";

export interface Car {
  id: number;
  make: string;
  model: string;
  year: number;
  color: string;
  category: string;
  transmission: CarTransmission;
  fuelType: CarFuelType;
  seats: number;
  bags: number;
  pricePerDay: number;
  available: boolean;
  imageUrl: string | null;
  features: string[];
  description: string | null;
  createdAt: string;
}