# lib/api-zod/src/generated/types/createCarInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRON Car Rental API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCarInputFuelType } from "./createCarInputFuelType";
import type { CreateCarInputTransmission } from "./createCarInputTransmission";

export interface CreateCarInput {
  make: string;
  model: string;
  year: number;
  color: string;
  category: string;
  transmission: CreateCarInputTransmission;
  fuelType: CreateCarInputFuelType;
  seats: number;
  bags: number;
  pricePerDay: number;
  available: boolean;
  imageUrl?: string | null;
  features?: string[];
  description?: string | null;
}