# lib/api-zod/src/generated/types/location.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRON Car Rental API
 * OpenAPI spec version: 0.1.0
 */

export interface Location {
  id: number;
  name: string;
  address: string;
  city: string;
  country: string;
  isAirport: boolean;
  active: boolean;
}